:recycle: Refactor submission models
